Move RSpec matchers to the lib directory
