+++ lib/manageiq/password/rspec_matchers.rb
@@ -1,4 +1,3 @@
-require "manageiq-password"
-require_relative "../../../spec/support/custom_matchers/be_decrypted"
-require_relative "../../../spec/support/custom_matchers/be_encrypted"
-require_relative "../../../spec/support/custom_matchers/be_encrypted_version"
+require "manageiq/password/rspec_matchers/be_decrypted"
+require "manageiq/password/rspec_matchers/be_encrypted"
+require "manageiq/password/rspec_matchers/be_encrypted_version"
