+++ lib/manageiq/password/rspec_matchers/be_encrypted.rb
@@ -1,3 +1,5 @@
+require "manageiq-password"
+
 RSpec::Matchers.define :be_encrypted do |expected|
   match do |actual|
     ManageIQ::Password.encrypted?(actual) && (
